refactoring bug fixes and further tables, time/date refactoring
